Preserve input dtype in to() and convolution-based operations
Several operations were silently upcasting float32 cubes to float64,
doubling memory usage:

- SpectralCube.to() / VaryingResolutionSpectralCube.to() /
  LowerDimensionalObject.to() multiply the data by a unit-conversion
  factor that bunit_converters always returns as float64, which
  upcasts float32 data on multiplication. Now the result is cast back
  to the original dtype.
- _apply_function_parallel_base (shared by apply_function_parallel_spatial
  and apply_function_parallel_spectral, and therefore by convolve_to,
  spatial_smooth, spectral_smooth, etc.) always allocated its output
  buffer as dtype=float (i.e. float64) regardless of the input cube's
  dtype. Now it matches the input dtype.
- VaryingResolutionSpectralCube.convolve_to had the same float64 buffer
  allocation bug independently, since it doesn't go through
  apply_function_parallel_spatial.
- DaskSpectralCube.convolve_to's convfunc didn't downcast the result of
  convolve_fft (which always returns float64) back to the input dtype
  when convolve_fft was explicitly requested (the default, plain
  convolve, already preserved dtype).

Fixes #995

+++ spectral_cube/lower_dimensional_structures.py
@@ -179,7 +179,10 @@ class LowerDimensionalObject(u.Quantity, BaseNDClass, HeaderMixinClass):
         factor = cube_utils.bunit_converters(self, unit, equivalencies=equivalencies,
                                              freq=freq)
 
-        converted_array = (self.quantity * factor).value
+        # the unit conversion factor is computed in double precision, which
+        # would otherwise silently inflate, e.g., float32 data to float64
+        # (see #995)
+        converted_array = (self.quantity * factor).value.astype(self.dtype, copy=False)
 
         # use private versions of variables, not the generated property
         # versions
